src/requirements.txt
bdfparser==2.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
flask==3.1.0
flask-cors==5.0.0
gtfs-realtime-bindings==2.0.0
idna==3.11
libsixel-python==0.5.0
numpy==2.3.5
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation==0.49b0
opentelemetry-instrumentation-flask==0.49b0
opentelemetry-instrumentation-requests==0.49b0
opentelemetry-exporter-jaeger==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
pillow==12.0.0
protobuf==6.33.2
pygame==2.6.1
requests==2.32.5
RGBMatrixEmulator==0.14.1
tornado==6.5.2
urllib3==2.6.1

